move: allSelectedSelector 파일 selectedSelector 폴더 하위로 이동
