+++ src/store/selectors/selectedSelector/allSelectedSelector.ts
